requirements.txt
streamlit==1.35.0
onnxruntime==1.17.1
torch==2.2.2
torchvision==0.17.2
opencv-python==4.9.0.80
Pillow==10.3.0
numpy==1.26.4
